1379 - fix with respect to Kamil comment

+++ app/Models/Photo.php
@@ -285,7 +285,7 @@ class Photo extends Model implements HasRandomID
 			return null;
 		}
 		// We need to format the framerate (stored as focal) -> max 2 decimal digits
-		return $this->isVideo() ? (string) round(intval($focal), 2) : $focal;
+		return $this->isVideo() ? (string) round(floatval($focal), 2) : $focal;
 	}
 
 	/**
